requirements.txt
flask
requests
python-dotenv
redis
rq